Taylor Thomas.
I fixed the draw circle because I had to add a semi colon at the end. I
also changed the box using CSS by changing the background color with
"background-color". I changed the color of the border with "color".
Lastly I used margin and put the box in the center of the page.